requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.9.0.80
Pillow>=10.0.0
numpy>=1.24.0